chore: add core DB object sync with local PouchDB for Quantel media objects
